requirements.txt
streamlit>=1.20.0
opencv-python
scikit-image
Pillow
imutils
numpy
requests
pyngrok

